Fix enum reading and writing

+++ src/Amicitia.IO.UnitTests/Binary/BinaryObjectWriterTests.cs
@@ -7,7 +7,6 @@ using System.IO;
 using System.Numerics;
 using System.Text;
 using System.Linq;
-
 using Amicitia.IO.Streams;
 
 namespace Amicitia.IO.Binary.Tests
@@ -121,9 +120,10 @@ namespace Amicitia.IO.Binary.Tests
             using ( var reader = new BinaryObjectReader( stream, StreamOwnership.Retain, Endianness.Little ) )
             {
                 Assert.AreEqual( 420, reader.ReadValueOffset<int>() );
+                
                 var offset = reader.ReadOffset();
-                Assert.AreEqual( 16, offset );
-                Assert.AreEqual( 69, reader.ReadValueAtOffset<int>( offset ) );
+                Assert.AreEqual( 16, (long)offset );
+                Assert.AreEqual( 69, reader.ReadValueAtOffset<int>( (long)offset ) );
             }
         }
